backend/requirements.txt
fastapi
uvicorn[standard]
opensearch-py
sqlalchemy
psycopg2-binary
sqlalchemy-utils
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
